+++ internal/engine/stats/clickhouse.go
@@ -1,0 +1,55 @@
+//
+//  Copyright 2026 The InfiniFlow Authors. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+package clickhouse
+
+//
+//// GetDriver gets global ClickHouse client instance
+//func GetDriver() *Driver {
+//	return globalDriver
+//}
+//
+//type Driver struct {
+//	conn    driver.Conn
+//	config  *clickhouse.Options
+//	channel chan *common.ModelUsage
+//	ctx     context.Context
+//	cancel  context.CancelFunc
+//	wg      sync.WaitGroup
+//}
+//
+//func (d *Driver) Status() (map[string]interface{}, error) {
+//	stats := d.conn.Stats()
+//	version, err := d.conn.ServerVersion()
+//	if err != nil {
+//		return nil, fmt.Errorf("failed to get server version: %w", err)
+//	}
+//	return map[string]interface{}{
+//		"max_open_conns": stats.MaxOpenConns,
+//		"max_idle_conns": stats.MaxIdleConns,
+//		"open":           stats.Open,
+//		"idle":           stats.Idle,
+//		"server_version": version.String(),
+//	}, nil
+//}
+//
+//func (d *Driver) Ping(ctx context.Context) error {
+//	return d.conn.Ping(ctx)
+//}
+//
+//func Close() error {
+//	return globalDriver.conn.Close()
+//}
