+++ internal/server/server_ee.go
@@ -37,7 +37,7 @@ func newEEServer() *EEServer {
 	return &EEServer{}
 }
 
-func StartServer(ctx context.Context, serverName string) error {
+func StartServer(ctx context.Context, cancel context.CancelFunc, serverName string) error {
 	if serverEE == nil {
 		return errors.New("server EE is nil")
 	}
